move storage.ts into kanbanwave
