+++ src/app/utils/index.ts
@@ -1,4 +1,3 @@
 export { default as setRef } from './setRef';
 export * from './format';
 export * as DB from './DB';
-export * from './storage';
