+++ src/kanbanwave/index.ts
@@ -1,1 +1,2 @@
+export * from './storage';
 export * from './types';
